requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.1
django-allauth==65.8.0
django-reset-migrations==0.4.0
django-role-permissions==3.2.0
et_xmlfile==2.0.0
fonttools==4.57.0
html5lib==1.1
idna==3.10
jwt==1.3.1
lxml==5.4.0
oauthlib==3.2.2
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.2.1
pycparser==2.22
pydyf==0.11.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
pyphen==0.17.2
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.4.0
weasyprint==65.1
webencodings==0.5.1
xhtml2pdf==0.2.17
XlsxWriter==3.2.3
zopfli==0.2.3.post1
